fix: apply styler to core

+++ core/systems/org/_public.ex
@@ -1,40 +1,27 @@
 defmodule Systems.Org.Public do
+  @moduledoc false
   use Core, :public
+  use Systems.Org.Internals
+
   import Ecto.Query, warn: false
 
   alias Core.Repo
-  alias Systems.Account.User
   alias Ecto.Multi
-
-  use Systems.{
-    Org.Internals
-  }
-
-  alias Systems.{
-    Content
-  }
+  alias Systems.Account.User
+  alias Systems.Content
 
   def get_node([_ | _] = identifier, preload \\ []) do
-    from(node in Node,
-      where: node.identifier == ^identifier,
-      preload: ^preload
-    )
-    |> Repo.one()
+    Repo.one(from(node in Node, where: node.identifier == ^identifier, preload: ^preload))
   end
 
   def get_node!(id, preload \\ [])
 
   def get_node!([_ | _] = identifier, preload) do
-    from(node in Node,
-      where: node.identifier == ^identifier,
-      preload: ^preload
-    )
-    |> Repo.one!()
+    Repo.one!(from(node in Node, where: node.identifier == ^identifier, preload: ^preload))
   end
 
   def get_node!(id, preload) do
-    from(node in Node, preload: ^preload)
-    |> Repo.get!(id)
+    Repo.get!(from(node in Node, preload: ^preload), id)
   end
 
   def create_node!(type, identifier, short_name, full_name) do
@@ -57,13 +44,12 @@ defmodule Systems.Org.Public do
     |> Multi.run(:org, fn _, %{short_name: short_name, full_name: full_name} ->
       {
         :ok,
-        %{
+        Org.Public.create_node!(%{
           type: type,
           identifier: identifier,
           short_name_bundle: short_name,
           full_name_bundle: full_name
-        }
-        |> Org.Public.create_node!()
+        })
       }
     end)
     |> Repo.commit()
@@ -73,17 +59,13 @@ defmodule Systems.Org.Public do
     {short_name_bundle, attrs} = Map.pop(attrs, :short_name_bundle, nil)
     {full_name_bundle, attrs} = Map.pop(attrs, :full_name_bundle, nil)
 
-    Node.create(attrs, short_name_bundle, full_name_bundle)
+    attrs
+    |> Node.create(short_name_bundle, full_name_bundle)
     |> Repo.insert!()
   end
 
   def get_link(%Node{id: from_id}, %Node{id: to_id}, preload \\ []) do
-    from(link in Link,
-      where: link.from_id == ^from_id,
-      where: link.to_id == ^to_id,
-      preload: ^preload
-    )
-    |> Repo.one()
+    Repo.one(from(link in Link, where: link.from_id == ^from_id, where: link.to_id == ^to_id, preload: ^preload))
   end
 
   def create_link!(%Node{} = from, %Node{} = to) do
@@ -95,7 +77,8 @@ defmodule Systems.Org.Public do
   end
 
   def add_user([_ | _] = identifier, %User{} = user) do
-    get_node!(identifier)
+    identifier
+    |> get_node!()
     |> add_user(user)
   end
 
@@ -109,55 +92,57 @@ defmodule Systems.Org.Public do
   end
 
   def delete_user([_ | _] = identifier, %User{} = user) do
-    get_node!(identifier)
+    identifier
+    |> get_node!()
     |> delete_user(user)
   end
 
   def delete_user(%Node{} = node, %User{} = user) do
-    from(ua in UserAssociation,
-      where: ua.org_id == ^node.id,
-      where: ua.user_id == ^user.id
-    )
-    |> Repo.delete_all()
+    Repo.delete_all(from(ua in UserAssociation, where: ua.org_id == ^node.id, where: ua.user_id == ^user.id))
   end
 
   def list_nodes(preload) do
-    list_nodes_query(preload)
+    preload
+    |> list_nodes_query()
     |> Repo.all()
   end
 
   def list_nodes(_, preload)
 
   def list_nodes(%User{} = user, preload) do
-    list_nodes_query(user, preload)
+    user
+    |> list_nodes_query(preload)
     |> Repo.all()
   end
 
   def list_nodes(type, preload) when is_atom(type) do
-    list_nodes_query(type, preload)
+    type
+    |> list_nodes_query(preload)
     |> Repo.all()
   end
 
   def list_nodes(_, _, preload)
 
   def list_nodes(%User{} = user, type, preload) when is_atom(type) do
-    list_nodes_query(user, type, preload)
+    user
+    |> list_nodes_query(type, preload)
     |> Repo.all()
   end
 
   def list_nodes(type, identifier_template, preload) when is_list(identifier_template) do
-    list_nodes_query(type, identifier_template, preload)
+    type
+    |> list_nodes_query(identifier_template, preload)
     |> Repo.all()
   end
 
   def list_nodes(type, _, preload) do
-    list_nodes_query(type, preload)
+    type
+    |> list_nodes_query(preload)
     |> Repo.all()
   end
 
   defp list_nodes_query(preload) do
-    from(node in Node)
-    |> query_preload(preload)
+    query_preload(from(node in Node), preload)
   end
 
   defp list_nodes_query(%User{} = user, type, preload) when is_atom(type) do
@@ -177,21 +162,20 @@ defmodule Systems.Org.Public do
   end
 
   defp list_nodes_query(%User{} = user, preload) do
-    from(node in Node,
-      inner_join: ua in UserAssociation,
-      on: ua.org_id == node.id,
-      inner_join: u in User,
-      on: u.id == ua.user_id,
-      where: u.id == ^user.id
+    query_preload(
+      from(node in Node,
+        inner_join: ua in UserAssociation,
+        on: ua.org_id == node.id,
+        inner_join: u in User,
+        on: u.id == ua.user_id,
+        where: u.id == ^user.id
+      ),
+      preload
     )
-    |> query_preload(preload)
   end
 
   defp list_nodes_query(type, preload) when is_atom(type) do
-    from(node in Node,
-      where: node.type == ^type
-    )
-    |> query_preload(preload)
+    query_preload(from(node in Node, where: node.type == ^type), preload)
   end
 
   defp query_preload(query, preload) when is_list(preload) do
